Moved Kodi installer/updater to a different screen
