+++ FireStarterProject/FireStarter/src/main/java/de/belu/firestarter/gui/AppActivity.java
@@ -44,6 +44,9 @@ public class AppActivity extends CustomFragment
      */
     SettingsProvider mSettings = SettingsProvider.getInstance(this.getActivity());
 
+    /** Latest FireStarter version found */
+    public static String LATEST_APP_VERSION = null;
+
     /**
      * Holds the gridview resource
      */
@@ -97,11 +100,11 @@ public class AppActivity extends CustomFragment
                     }
 
                     // Check for new update
-                    if(!mSettings.getHaveUpdateSeen() && FireStarterUpdater.isVersionNewer(Tools.getCurrentAppVersion(getActivity()), FireStarterUpdater.LATEST_VERSION))
+                    if(!mSettings.getHaveUpdateSeen() && FireStarterUpdater.isVersionNewer(Tools.getCurrentAppVersion(getActivity()), LATEST_APP_VERSION))
                     {
                         AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
 
-                        builder.setTitle("FireStarter " + FireStarterUpdater.LATEST_VERSION);
+                        builder.setTitle("FireStarter " + LATEST_APP_VERSION);
                         builder.setMessage("There is a new version of FireStarter, do you want to update?");
 
                         builder.setPositiveButton("YES", new DialogInterface.OnClickListener()
