+++ FireStarterProject/FireStarter/src/main/java/de/belu/firestarter/gui/LeftBarItemsListAdapter.java
@@ -56,6 +56,7 @@ public class LeftBarItemsListAdapter extends BaseAdapter
             mItems.add(new FragmentListItem(mContext.getResources().getString(R.string.leftbar_infos), InfosPrefActivity.class.getName()));
             mItems.add(new FragmentListItem(mContext.getResources().getString(R.string.leftbar_updates), UpdateActivity.class.getName()));
             mItems.add(new FragmentListItem(mContext.getResources().getString(R.string.leftbar_settings), PreferenceActivity.class.getName()));
+            mItems.add(new FragmentListItem(mContext.getResources().getString(R.string.leftbar_install), InstallActivity.class.getName()));
         }
     }
